refactoring / cleanup
