+++ simapp/testapp.go
@@ -2,6 +2,7 @@ package simapp
 
 import (
 	"encoding/json"
+	"github.com/NibiruChain/nibiru/app"
 	"os"
 	"path/filepath"
 	"time"
@@ -15,7 +16,6 @@ import (
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	tmdb "github.com/tendermint/tm-db"
 
-	"github.com/NibiruChain/nibiru/app/wasmconfig"
 	"github.com/NibiruChain/nibiru/x/common"
 	pricefeedtypes "github.com/NibiruChain/nibiru/x/pricefeed/types"
 )
@@ -66,7 +66,7 @@ func NewTestNibiruAppWithGenesis(gen GenesisState) *NibiruTestApp {
 		/*invCheckPeriod=*/ 0,
 		/*encodingConfig=*/ encoding,
 		/*appOpts=*/ simapp.EmptyAppOptions{},
-		wasmconfig.DefaultConfig(),
+		app.DefaultWasmConfig(),
 	)
 
 	stateBytes, err := json.MarshalIndent(gen, "", " ")
